process all columns and check data type with regex

+++ lab1/MapReduceJavaExample/src/main/java/io/datamass/ReducerWordCount.java
@@ -1,22 +1,25 @@
 package io.datamass;
 
 
+import static io.datamass.MapWordCount.TOTAL;
+
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.mapreduce.Reducer;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import java.io.IOException;
 
 
-public class ReducerWordCount extends Reducer<Text, IntWritable, Text, IntWritable>
+public class ReducerWordCount extends Reducer<Text, DoubleWritable, Text, DoubleWritable>
 {
-    public void reduce(Text word, Iterable<IntWritable> values, Context con) throws IOException, InterruptedException
+    public void reduce(Text word, Iterable<DoubleWritable> values, Context con) throws IOException, InterruptedException
     {
-        int sum = 0;
-        for(IntWritable value : values)
+        double sum = 0;
+        for(DoubleWritable value : values)
         {
             sum += value.get();
         }
-        con.write(word, new IntWritable(sum));
+        con.write(word, new DoubleWritable(sum));
     }
 }
